+++ metadata.rb
@@ -1,15 +1,17 @@
 # encoding: UTF-8
-# rubocop:disable SingleSpaceBeforeFirstArg
 name             'wildfly'
 maintainer       'Brian Dwyer - Intelligent Digital Services'
 maintainer_email 'bdwyertech'
 license          'Apache License, Version 2.0'
 description      'Installs/Configures wildfly'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.2.0'
+version          '0.3.1'
 
 supports 'centos'
 
 depends 'apt'
 depends 'yum'
 depends 'java', '~> 1.22'
+
+source_url 'https://github.com/bdwyertech/chef-wildfly' if respond_to?(:source_url)
+issues_url 'https://github.com/bdwyertech/chef-wildfly/issues' if respond_to?(:issues_url)