+++ providers/datasource.rb
@@ -19,6 +19,8 @@
 require 'etc'
 include Chef::Mixin::ShellOut
 
+use_inline_resources
+
 # Support whyrun
 def whyrun_supported?
   true
@@ -50,33 +52,41 @@ def load_current_resource
   @current_resource.jndiname(@new_resource.jndiname)
   @current_resource.drivername(@new_resource.drivername)
   @current_resource.connectionurl(@new_resource.connectionurl)
+  @current_resource.username(@new_resource.username)
+  @current_resource.password(@new_resource.password)
+  @current_resource.sensitive(@new_resource.sensitive)
   @current_resource.exists = true if datasource_exists?(@current_resource.name)
   # TODO: Set @current_resource port properties from command output
 end
 
-private
-
 def datasource_exists?(name)
-  result = shell_out("su #{node['wildfly']['user']} -s /bin/bash -c \"#{node['wildfly']['base']}/bin/jboss-cli.sh -c ' /subsystem=datasources/data-source=#{name}:read-resource'\"")
+  result = shell_out("su #{node['wildfly']['user']} -s /bin/bash -c \"#{node['wildfly']['base']}/bin/jboss-cli.sh -c ' /subsystem=datasources/data-source=#{name.gsub('/', '\/')}:read-resource'\"")
   result.exitstatus == 0
 end
 
+private
+
 def create_datasource
-  bash 'install_datasource' do
+  params = %W(--name=#{new_resource.name} --jndi-name=#{new_resource.jndiname} --driver-name=#{new_resource.drivername} --connection-url=#{new_resource.connectionurl})
+  params << "--user-name=#{new_resource.username}" if new_resource.username
+  params << "--password=#{new_resource.password}" if new_resource.password
+
+  bash "install_datasource #{new_resource.name}" do
     user node['wildfly']['user']
     cwd node['wildfly']['base']
-    code <<-EOH
-      bin/jboss-cli.sh -c command="data-source add --name=#{new_resource.name} --jndi-name=#{new_resource.jndiname} --driver-name=#{new_resource.drivername} --connection-url=#{new_resource.connectionurl}"
-    EOH
+    code "bin/jboss-cli.sh -c command=\"data-source add #{params.join(' ')}\""
+    sensitive new_resource.sensitive
+    not_if { datasource_exists?(new_resource.name) }
   end
 end
 
 def delete_datasource
-  bash 'remove_datasource' do
+  bash "remove_datasource #{new_resource.name}" do
     user node['wildfly']['user']
     cwd node['wildfly']['base']
     code <<-EOH
       bin/jboss-cli.sh -c command="data-source remove --name=#{new_resource.name}"
     EOH
+    only_if { datasource_exists?(new_resource.name) }
   end
 end