+++ resources/datasource.rb
@@ -24,5 +24,14 @@ attribute :name,          kind_of: String, required: true, name_attribute: true
 attribute :jndiname,      kind_of: String
 attribute :drivername,    kind_of: String
 attribute :connectionurl, kind_of: String
+attribute :username, kind_of: [String, NilClass]
+attribute :password, kind_of: [String, NilClass]
+attribute :sensitive, kind_of: [TrueClass, FalseClass] # , default: true - see initialize below
 
 attr_accessor :exists
+
+# Chef will override sensitive back to its global value, so set default to true in init
+def initialize(*args)
+  super
+  @sensitive = true
+end