+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 WildFly Cookbook CHANGELOG
 ==========================
 
+# 0.3.1 (2016-04-11)
+* Chef will override sensitive back to its global value
+
+# 0.3.0 (2016-02-24)
+* Add sensitive and optional username/password parameters to datasource resource.
+* Fix missing jboss.jdbc-driver.mysql by restarting Wildfly immediately after installing mysql_connector.
+* Add ChefSpec Matchers.
+
 # 0.2.0 (2015-09-09)
 * Make install of Java optional.